requirements.txt
ipykernel
requests
beautifulsoup4
lxml-html-clean  
nltk
scikit-learn
spacy
dotenv
google-genai
uvicorn
fastapi
gtts
googletrans